refactor: Split middleware module into separate files for better organization refactor: Simplify JWT middleware to skip authentication for Swagger UI routes
